move sam to `contrib.models`
