+++ kornia/contrib/models/sam/__init__.py
@@ -1,0 +1,1 @@
+from kornia.contrib.models.sam.model import Sam, SamConfig, SamModelType
