+++ kornia/contrib/models/sam/architecture/common.py
