+++ kornia/contrib/models/sam/architecture/transformer.py
@@ -9,7 +9,7 @@ import math
 
 from torch import nn
 
-from kornia.contrib.sam.architecture.common import MLPBlock
+from kornia.contrib.models.sam.architecture.common import MLPBlock
 from kornia.core import Module, Tensor
 from kornia.core.check import KORNIA_CHECK
