--- kornia/contrib/sam/__init__.py
@@ -1,1 +1,0 @@
-from kornia.contrib.sam.model import Sam, SamConfig, SamModelType
